Fixed kickoffs, punts and returns
Removed legacy files / functions. Worked in updates to the calculations for kick returns for increased accuracy. This includes the frequency of TD's. Punt returns were also reviewed and only made slight adjustments. Frequency of TDs on kick returns is still under review (it is suspected it may have been higher in the past)